+++ src/vite-env.d.ts
@@ -12,3 +12,8 @@ declare module '*.glb?url' {
   const url: string;
   export default url;
 }
+
+declare module '*.png' {
+  const url: string;
+  export default url;
+}